tests/wpt/tests/css/css-fonts/parsing/font-kerning-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-kerning with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-kerning-prop">
<meta name="assert" content="font-kerning supports only the grammar 'auto | normal | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-kerning', 'normal auto');
test_invalid_value('font-kerning', 'none, auto');
</script>
</body>
</html>
